model for widget grid, selector & chart
- create new model for chart
- create save, delete, get data on widget grid model
